+++ packages/engine-multi/src/worker/thread/runtime.ts
@@ -4,13 +4,19 @@ import {
   ENGINE_REJECT_TASK,
   ENGINE_RESOLVE_TASK,
   ENGINE_RUN_TASK,
+  type WorkerEvents,
 } from '../events';
 import ensurePayloadSize from '../../util/ensure-payload-size';
 
 // This constrains the size of any payloads coming out of the worker thread
 // Payloads exceeding this will be redacted
 // (in practice I think this is specifically the state and message keys)
-let payloadLimitMb: number | undefined;
+export type PayloadLimits = {
+  default?: number;
+  [key: string]: number | undefined;
+} & Partial<Record<WorkerEvents, number>>;
+
+let payloadLimits: PayloadLimits | undefined;
 
 type TaskRegistry = Record<string, (...args: any[]) => Promise<any>>;
 
@@ -38,27 +44,43 @@ type Event = {
 
 type Options = {
   memoryLimitMb?: number;
-  payloadLimitMb?: number;
+  payloadLimits?: PayloadLimits;
 };
 
-export const publish = (
+export const publish = async (
   type: string,
   payload: Omit<Event, 'threadId' | 'type'>
 ) => {
   // Validate the size of every outgoing message
   // Redact any payloads that are too large
-  const safePayload = ensurePayloadSize(payload, payloadLimitMb);
-  const x = {
+  const limit =
+    payloadLimits?.[type as keyof PayloadLimits] ?? payloadLimits?.default;
+  const safePayload = await ensurePayloadSize(payload, limit);
+
+  parentPort!.postMessage({
     type,
     threadId,
     processId,
     ...safePayload,
-  };
-  parentPort!.postMessage(x);
+    ...payload,
+  });
+};
+
+// publishes without payload validation
+export const publishSync = async (
+  type: string,
+  payload: Omit<Event, 'threadId' | 'type'>
+) => {
+  parentPort!.postMessage({
+    type,
+    threadId,
+    processId,
+    ...payload,
+  });
 };
 
 const run = (task: string, args: any[], options: Options = {}) => {
-  payloadLimitMb = options.payloadLimitMb;
+  payloadLimits = options.payloadLimits;
   tasks[task](...args)
     .then((result) => {
       publish(ENGINE_RESOLVE_TASK, {
@@ -76,12 +98,12 @@ const run = (task: string, args: any[], options: Options = {}) => {
       });
     })
     .finally(() => {
-      payloadLimitMb = undefined;
+      payloadLimits = undefined;
     });
 };
 
 process.on('exit', (code) => {
-  publish(ENGINE_REJECT_TASK, {
+  publishSync(ENGINE_REJECT_TASK, {
     error: {
       name: 'ExitError',
       message: `Worker thread exited with code: ${code}`,